# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EcoSphere ESG Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/environmental:
    get:
      operationId: getEnvironmentalReport
      tags: [reports]
      parameters:
        - { name: departmentId, in: query, schema: { type: integer } }
        - { name: startDate, in: query, schema: { type: string, format: date } }
        - { name: endDate, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Environmental report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentalReport"
  /reports/social:
    get:
      operationId: getSocialReport
      tags: [reports]
      parameters:
        - { name: departmentId, in: query, schema: { type: integer } }
        - { name: startDate, in: query, schema: { type: string, format: date } }
        - { name: endDate, in: query, schema: { type: string, format: date } }
        - { name: employeeId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Social report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialReport"
  /reports/governance:
    get:
      operationId: getGovernanceReport
      tags: [reports]
      parameters:
        - { name: departmentId, in: query, schema: { type: integer } }
        - { name: startDate, in: query, schema: { type: string, format: date } }
        - { name: endDate, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: Governance report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceReport"
  /reports/esg-summary:
    get:
      operationId: getEsgSummaryReport
      tags: [reports]
      parameters:
        - { name: departmentId, in: query, schema: { type: integer } }
        - { name: startDate, in: query, schema: { type: string, format: date } }
        - { name: endDate, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: ESG summary report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EsgSummaryReport"
components:
  schemas:
...
    CarbonTransaction:
      type: object
      required: [id, departmentId, emissionFactorId, quantity, totalEmission, transactionDate, source]
      properties:
        id: { type: integer }
        departmentId: { type: integer }
        departmentName: { type: ["string", "null"] }
        emissionFactorId: { type: integer }
        emissionFactorName: { type: ["string", "null"] }
        quantity: { type: number }
        totalEmission: { type: number }
        source: { type: string, enum: [purchase, manufacturing, expense, fleet, manual] }
        description: { type: ["string", "null"] }
        transactionDate: { type: string, format: date }
        createdAt: { type: string, format: date-time }
...
    EnvironmentalGoal:
      type: object
      required: [id, title, targetValue, currentValue, unit, deadline, status]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        departmentId: { type: ["integer", "null"] }
        departmentName: { type: ["string", "null"] }
        targetValue: { type: number }
        currentValue: { type: number }
        unit: { type: string }
        deadline: { type: string, format: date }
        status: { type: string, enum: [active, achieved, missed, cancelled] }
        progressPercent: { type: number }
...
    EmissionByDept:
      type: object
      required: [departmentId, departmentName, totalEmission]
      properties:
        departmentId: { type: integer }
        departmentName: { type: string }
        totalEmission: { type: number }
    EmissionBySource:
      type: object
      required: [source, totalEmission]
      properties:
        source: { type: string }
        totalEmission: { type: number }
...
    GenderDist:
      type: object
      required: [gender, count, percent]
      properties:
        gender: { type: string }
        count: { type: integer }
        percent: { type: number }
    TopParticipant:
      type: object
      required: [employeeId, employeeName, participationCount, pointsEarned]
      properties:
        employeeId: { type: integer }
        employeeName: { type: string }
        participationCount: { type: integer }
        pointsEarned: { type: integer }
...
    IssuesBySeverity:
      type: object
      required: [severity, count]
      properties:
        severity: { type: string }
        count: { type: integer }
    DepartmentRanking:
      type: object
      required: [rank, departmentId, departmentName, totalScore]
      properties:
        rank: { type: integer }
        departmentId: { type: integer }
        departmentName: { type: string }
        environmentalScore: { type: number }
        socialScore: { type: number }
        governanceScore: { type: number }
        totalScore: { type: number }

    # ── Report Schemas ───────────────────────────────────────────────────────
    EnvironmentalReport:
      type: object
      properties:
        generatedAt: { type: string, format: date-time }
        period: { $ref: "#/components/schemas/ReportPeriod" }
        totalEmissions: { type: number }
        emissionsByDepartment: { type: array, items: { $ref: "#/components/schemas/EmissionByDept" } }
        emissionsBySource: { type: array, items: { $ref: "#/components/schemas/EmissionBySource" } }
        goalsStatus: { type: array, items: { $ref: "#/components/schemas/EnvironmentalGoal" } }
        transactions: { type: array, items: { $ref: "#/components/schemas/CarbonTransaction" } }
    SocialReport:
      type: object
      properties:
        generatedAt: { type: string, format: date-time }
        period: { $ref: "#/components/schemas/ReportPeriod" }
        totalParticipations: { type: integer }
        approvalRate: { type: number }
        activitiesByCategory: { type: array, items: { $ref: "#/components/schemas/ActivityByCategory" } }
        genderDistribution: { type: array, items: { $ref: "#/components/schemas/GenderDist" } }
        topParticipants: { type: array, items: { $ref: "#/components/schemas/TopParticipant" } }
    ActivityByCategory:
      type: object
      required: [categoryName, count]
      properties:
        categoryName: { type: string }
        count: { type: integer }
    GovernanceReport:
      type: object
      properties:
        generatedAt: { type: string, format: date-time }
        period: { $ref: "#/components/schemas/ReportPeriod" }
        activePolicies: { type: integer }
        acknowledgementRate: { type: number }
        auditsCompleted: { type: integer }
        complianceIssuesRaised: { type: integer }
        complianceIssuesResolved: { type: integer }
        issuesBySeverity: { type: array, items: { $ref: "#/components/schemas/IssuesBySeverity" } }
    EsgSummaryReport:
      type: object
      properties:
        generatedAt: { type: string, format: date-time }
        period: { $ref: "#/components/schemas/ReportPeriod" }
        overallScore: { type: number }
        environmentalScore: { type: number }
        socialScore: { type: number }
        governanceScore: { type: number }
        departmentRankings: { type: array, items: { $ref: "#/components/schemas/DepartmentRanking" } }
        highlights: { type: array, items: { type: string } }
    ReportPeriod:
      type: object
      required: [startDate, endDate]
      properties:
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }